fix: add in plumbing for exporting calc compound rate
