+++ packages/math-utils/src/formatters/compounded-interest/index.ts
@@ -1,0 +1,2 @@
+export * from './calculate-compounded-interest';
+export * from './types';
