+++ packages/math-utils/src/formatters/compounded-interest/types.ts
@@ -1,0 +1,12 @@
+import { BigNumberValue } from '../../bignumber';
+
+export interface CalculateCompoundedInterestRateRequest {
+  rate: BigNumberValue;
+  currentTimestamp: number;
+  lastUpdateTimestamp: number;
+}
+
+export interface CalculateCompoundedRateRequest {
+  rate: BigNumberValue;
+  duration: BigNumberValue;
+}
