+++ packages/math-utils/src/formatters/reserve/calculate-reserve-debt.ts
@@ -1,6 +1,6 @@
 import BigNumber from 'bignumber.js';
 import { rayMul } from '../../ray.math';
-import { calculateCompoundedInterest } from '../compounded-interest/calculate-compounded-interest';
+import { calculateCompoundedInterestRate } from '../compounded-interest/calculate-compounded-interest';
 
 export interface CalculateReserveDebtRequest {
   totalScaledVariableDebt: string;
@@ -45,7 +45,7 @@ function getTotalVariableDebt(
       reserveDebt.totalScaledVariableDebt,
       reserveDebt.variableBorrowIndex,
     ),
-    calculateCompoundedInterest({
+    calculateCompoundedInterestRate({
       rate: reserveDebt.variableBorrowRate,
       currentTimestamp,
       lastUpdateTimestamp: reserveDebt.lastUpdateTimestamp,
@@ -59,7 +59,7 @@ function getTotalStableDebt(
 ): BigNumber {
   return rayMul(
     reserveDebt.totalPrincipalStableDebt,
-    calculateCompoundedInterest({
+    calculateCompoundedInterestRate({
       rate: reserveDebt.averageStableRate,
       currentTimestamp,
       lastUpdateTimestamp: reserveDebt.stableDebtLastUpdateTimestamp,
